+++ src/main/java/com/todoay/api/domain/todo/controller/DueDateTodoController.java
@@ -1,6 +1,7 @@
 package com.todoay.api.domain.todo.controller;
 
-import com.todoay.api.domain.todo.dto.*;
+import com.todoay.api.domain.todo.dto.daily.DailyTodoSaveResponseDto;
+import com.todoay.api.domain.todo.dto.duedate.*;
 import com.todoay.api.domain.todo.service.DueDateTodoCRUDService;
 import com.todoay.api.global.exception.ErrorResponse;
 import com.todoay.api.global.exception.ValidErrorResponse;
