+++ src/main/java/com/todoay/api/domain/todo/dto/daily/DailyTodoModifyRequestDto.java
@@ -1,4 +1,4 @@
-package com.todoay.api.domain.todo.dto;
+package com.todoay.api.domain.todo.dto.daily;
 
 import com.todoay.api.domain.hashtag.dto.HashtagInfoDto;
 import lombok.Builder;
