+++ src/main/java/com/todoay/api/domain/todo/dto/duedate/OrderCondition.java
@@ -1,4 +1,4 @@
-package com.todoay.api.domain.todo.dto;
+package com.todoay.api.domain.todo.dto.duedate;
 
 public enum OrderCondition {
     DUEDATE,IMPORTANCE
