+++ src/main/java/com/todoay/api/domain/todo/service/DueDateTodoCRUDService.java
@@ -1,6 +1,6 @@
 package com.todoay.api.domain.todo.service;
 
-import com.todoay.api.domain.todo.dto.*;
+import com.todoay.api.domain.todo.dto.duedate.*;
 
 import java.util.List;
