+++ src/test/java/com/todoay/api/domain/todo/service/DailyTodoCRUDServiceImplTest.java
@@ -6,10 +6,10 @@ import com.todoay.api.domain.category.entity.Category;
 import com.todoay.api.domain.category.repository.CategoryRepository;
 import com.todoay.api.domain.hashtag.entity.Hashtag;
 import com.todoay.api.domain.hashtag.repository.HashtagRepository;
-import com.todoay.api.domain.todo.dto.DailyTodoDailyDateModifyRequestDto;
-import com.todoay.api.domain.todo.dto.DailyTodoReadDetailResponseDto;
-import com.todoay.api.domain.todo.dto.DailyTodoReadResponseDto;
-import com.todoay.api.domain.todo.dto.DailyTodoRepeatRequestDto;
+import com.todoay.api.domain.todo.dto.daily.DailyTodoDailyDateModifyRequestDto;
+import com.todoay.api.domain.todo.dto.daily.DailyTodoReadDetailResponseDto;
+import com.todoay.api.domain.todo.dto.daily.DailyTodoReadResponseDto;
+import com.todoay.api.domain.todo.dto.daily.DailyTodoRepeatRequestDto;
 import com.todoay.api.domain.todo.entity.DailyTodo;
 import com.todoay.api.domain.todo.repository.DailyTodoRepository;
 import org.junit.jupiter.api.AfterEach;
